requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
trafilatura>=1.6.1,<1.7.0
openai>=1.0.0
supabase>=2.0.0
pydantic>=1.8.0,<2.0.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
python-dotenv>=0.19.0
ratelimit>=2.2.1
tenacity>=8.2.0
